refactor(backend): move platform foundation packages to internal/core
Adopt the repo-wide core/platform/features naming convention: the Phase 1
foundation (config, database, httpserver, logging, jobs, inbox) is our own
infrastructure and moves to internal/core. internal/platform is reserved for
third-party SaaS adapters (added next). Import-path-only change; no behavior
change.
